+++ CHANGELOG.md
@@ -10,6 +10,7 @@ the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 ## [Unreleased]
 ### Added
 ### Changed
+- `get_os()` raises `RuntimeError` if the operating system is not supported
 ### Fixed
 ### Removed
 ### Security